moved app specific stuff to BlueshiftAttrApp class from BlueshiftAttrDevice class
