+++ android-sdk/src/main/java/com/blueshift/BlueshiftAttrApp.java
@@ -7,6 +7,11 @@ import android.os.Build;
 import android.support.v4.app.NotificationManagerCompat;
 
 import com.blueshift.util.BlueshiftUtils;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.FirebaseApp;
+import com.google.firebase.iid.FirebaseInstanceId;
+import com.google.firebase.iid.InstanceIdResult;
 
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -37,6 +42,85 @@ public class BlueshiftAttrApp extends JSONObject {
         addAppVersion(context);
         addInAppEnabledStatus(context);
         addPushEnabledStatus(context);
+        addFirebaseInstanceId(context);
+        addFirebaseToken(context);
+    }
+
+    private void addFirebaseToken(Context context) {
+        if (!BlueshiftUtils.isPushEnabled(context)) return;
+
+        try {
+            addFirebaseToken();
+        } catch (Exception e) {
+            // tickets#8919 reported an issue with fcm token fetch. this is the
+            // fix for the same. we are manually calling initializeApp and trying
+            // to get token again.
+            FirebaseApp.initializeApp(context);
+            try {
+                addFirebaseToken();
+            } catch (Exception e1) {
+                BlueshiftLogger.e(TAG, e1);
+            }
+        }
+    }
+
+    private void addFirebaseToken() {
+        Task<InstanceIdResult> task = FirebaseInstanceId.getInstance().getInstanceId();
+        task.addOnSuccessListener(new OnSuccessListener<InstanceIdResult>() {
+            @Override
+            public void onSuccess(InstanceIdResult instanceIdResult) {
+                String deviceToken = instanceIdResult.getToken();
+                setFirebaseToken(deviceToken);
+            }
+        });
+    }
+
+    private void setFirebaseToken(String firebaseToken) {
+        synchronized (instance) {
+            try {
+                instance.put(BlueshiftConstants.KEY_DEVICE_TOKEN, firebaseToken);
+            } catch (JSONException e) {
+                BlueshiftLogger.e(TAG, e);
+            }
+        }
+    }
+
+    private void addFirebaseInstanceId(Context context) {
+        try {
+            addFirebaseInstanceId();
+        } catch (Exception e) {
+            // tickets#8919 reported an issue with fcm token fetch. this is the
+            // fix for the same. we are manually calling initializeApp and trying
+            // to get token again.
+            FirebaseApp.initializeApp(context);
+
+            try {
+                addFirebaseInstanceId();
+            } catch (Exception ex) {
+                BlueshiftLogger.e(TAG, ex);
+            }
+        }
+    }
+
+    private void addFirebaseInstanceId() {
+        String instanceId = FirebaseInstanceId.getInstance().getId();
+        setFirebaseInstanceId(instanceId);
+    }
+
+    private void setFirebaseInstanceId(String instanceId) {
+        synchronized (instance) {
+            try {
+                instance.putOpt(BlueshiftConstants.KEY_FIREBASE_INSTANCE_ID, instanceId);
+            } catch (JSONException e) {
+                BlueshiftLogger.e(TAG, e);
+            }
+        }
+    }
+
+    public void updateFirebaseToken(String newToken) {
+        if (newToken != null) {
+            setFirebaseToken(newToken);
+        }
     }
 
     private void addPushEnabledStatus(Context context) {
